requirements.txt
numpy==1.18.1
pandas==0.25.3

matplotlib==3.1.2
seaborn==0.9.0
umap-learn==0.3.10

torch==1.2.0

networkx==2.4
python-igraph==0.8.3

rpy2==3.4.5

# Below are not used
# auto_encoders.gae.train
# munkres==1.1.2
# tqdm==4.48.0